Split data classes (Community, GardenReservation, MedianTable, SimplifiedMyPageCommunity, SimplifiedCommunity)
