+++ app/src/main/java/com/kangdroid/k_labapplication/data/MedianTable.kt
@@ -1,0 +1,7 @@
+package com.kangdroid.k_labapplication.data
+
+data class MedianTable(
+    var userId: Long, // [User PK]
+    var communityId: Long, // [Community PK],
+    var isRequestConfirmed: Boolean
+)
